+++ server/database/models/Ravitaillement_detail.model.js
@@ -4,10 +4,10 @@ import db from "../../config/Database.js";
 const Ravitaillement_detail = db.define(
   "ravitaillement_detail",
   {
-    prix_unit: { type: DataTypes.DOUBLE(10), allowNull: false },
-    prix_ht: { type: DataTypes.DOUBLE(10), allowNull: false },
-    quantite_demande: { type: DataTypes.DOUBLE(10), allowNull: false },
-    quantite_livraison: { type: DataTypes.DOUBLE(10), allowNull: false },
+    prix_unit: { type: DataTypes.DOUBLE, allowNull: false },
+    prix_ht: { type: DataTypes.DOUBLE, allowNull: false },
+    quantite_demande: { type: DataTypes.DOUBLE, allowNull: false },
+    quantite_livraison: { type: DataTypes.DOUBLE, allowNull: false },
     // unite_achat: { type: DataTypes.INTEGER(10), allowNull: false },
     status: { type: DataTypes.BOOLEAN, defaultValue: "1" },
   },
