+++ server/database/models/Vente_detail.model.js
@@ -3,9 +3,9 @@ import db from "../../config/Database.js";
 const Vente_detail = db.define(
   "vente_detail",
   {
-    quantite_vente: { type: DataTypes.DOUBLE(10), allowNull: false },
-    prix_vente: { type: DataTypes.DOUBLE(10), allowNull: false },
-    montant_vente: { type: DataTypes.DOUBLE(10), allowNull: false },
+    quantite_vente: { type: DataTypes.DOUBLE, allowNull: false },
+    prix_vente: { type: DataTypes.DOUBLE, allowNull: false },
+    montant_vente: { type: DataTypes.DOUBLE, allowNull: false },
     status: { type: DataTypes.BOOLEAN, defaultValue: "1" },
   },
   { paranoid: true, freezeTableName: true }
